Fix for Collect+Lookup bug.
